move examples section out of stdlib pyweb to examples.py in the demo itself
