+++ pyscript.core/test/ui/pyscript.toml
@@ -1,1 +1,4 @@
 packages = []
+
+[files]
+"./examples.py" = "./examples.py"
